+++ Source/AnimusExMachina/Private/Config/AxMConfig.cpp
@@ -1,0 +1,3 @@
+// Copyright ArconicLabs. All Rights Reserved.
+
+#include "Config/AxMConfig.h"